Renamed the 'main' package to the 'gui' package
