+++ src/main/java/brickdestroy/elements/Game.java
@@ -9,7 +9,7 @@ import java.awt.geom.Point2D;
 import java.awt.geom.Rectangle2D;
 import java.util.Random;
 
-import brickdestroy.main.MainFrame;
+import brickdestroy.gui.MainFrame;
 
 public class Game {
