+++ src/main/java/brickdestroy/gui/GameController.java
@@ -1,4 +1,4 @@
-package brickdestroy.main;
+package brickdestroy.gui;
 
 import javax.swing.JPanel;
 import javax.swing.Timer;
